requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
dnspython==2.4.2
genson==1.2.2
numpy==1.26.2
pandas==2.1.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pycparser==2.21
pymongo==4.6.0
pypdfium2==4.24.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
